requirements.txt
facebook_sdk==3.0.0a0
Flask
numpy
requests
selenium
wget
beautifulsoup4
